shift player interact to own class
